team_dogwood/coursework_two/pyproject.toml
[project]
name = "team-dogwood-coursework-two"
version = "0.1.0"
description = ""
authors = [
    {name = "Team Dogwood"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "black (>=24.10.0,<25.0.0)",
    "isort (>=6.0.0,<7.0.0)",
    "flake8 (>=7.1.1,<8.0.0)",
    "bandit (==1.8.2)",
    "pytest (>=8.3.4,<9.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "safety (>=3.3.1,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "minio (>=7.2.15,<8.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sphinx-wagtail-theme (>=6.5.0,<7.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "llama-cloud-services (>=0.6.12,<0.7.0)",
    "llama-index-core (>=0.12.30,<0.13.0)",
    "llama-index-readers-file (>=0.4.7,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "llama-index-llms-openai (>=0.3.33,<0.4.0)",
    "llama-index-embeddings-openai (>=0.3.1,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
